+++ src/basic/hooks/useProductFormHandlers.ts
@@ -1,0 +1,142 @@
+import { Dispatch, SetStateAction, useCallback } from "react";
+import { NewProductForm } from "../types/product";
+import { useNotification } from "./useNotification";
+import { DISCOUNT_OPTIONS, INITIAL_PRODUCT_FORM, processNumericInput, productValidation } from "../utils/productUtils";
+
+interface IUseProductFormHandlersProps {
+  setProductForm: Dispatch<SetStateAction<NewProductForm>>;
+  productForm: NewProductForm;
+  setEditingProduct: Dispatch<SetStateAction<string | null>>;
+  setShowProductForm: Dispatch<SetStateAction<boolean>>;
+}
+
+export const useProductFormHandlers = ({ setProductForm, productForm, setEditingProduct, setShowProductForm }: IUseProductFormHandlersProps) => {
+  const { showToast } = useNotification();
+  /** 상품 폼 업데이트 */
+  const updateField = useCallback(
+    <K extends keyof NewProductForm>(field: K, value: NewProductForm[K]) => {
+      setProductForm((prev) => ({ ...prev, [field]: value }));
+    },
+    [setProductForm]
+  );
+
+  /** 상품명 입력 핸들러 */
+  const handleNameChange = useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      updateField("name", e.target.value);
+    },
+    [updateField]
+  );
+
+  /** 설명 입력 핸들러 */
+  const handleDescriptionChange = useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      updateField("description", e.target.value);
+    },
+    [updateField]
+  );
+
+  /** 가격 입력 (숫자만 허용) */
+  const handlePriceChange = useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      const processedValue = processNumericInput(e.target.value);
+      if (processedValue !== null) {
+        updateField("price", processedValue);
+      }
+    },
+    [updateField]
+  );
+
+  /** 가격 포커스 해제 시 검증 */
+  const handlePriceBlur = useCallback(
+    (e: React.FocusEvent<HTMLInputElement>) => {
+      const validation = productValidation.validatePrice(parseInt(e.target.value) || 0);
+
+      if (!validation.isValid) {
+        showToast(validation.message!, "error");
+        updateField("price", validation.correctedValue!);
+      }
+    },
+    [updateField, showToast]
+  );
+
+  /** 재고 입력 (숫자만 허용) */
+  const handleStockChange = useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      const processedValue = processNumericInput(e.target.value);
+      if (processedValue !== null) {
+        updateField("stock", processedValue);
+      }
+    },
+    [updateField]
+  );
+
+  /** 재고 포커스 해제 시 검증 */
+  const handleStockBlur = useCallback(
+    (e: React.FocusEvent<HTMLInputElement>) => {
+      const validation = productValidation.validateStock(parseInt(e.target.value) || 0);
+
+      if (!validation.isValid) {
+        showToast(validation.message!, "error");
+        updateField("stock", validation.correctedValue!);
+      }
+    },
+    [updateField, showToast]
+  );
+
+  /** 할인 수량 변경 */
+  const handleDiscountQuantityChange = useCallback(
+    (index: number, value: string) => {
+      const newDiscounts = [...productForm.discounts];
+      newDiscounts[index].quantity = parseInt(value) || 0;
+      updateField("discounts", newDiscounts);
+    },
+    [productForm.discounts, updateField]
+  );
+
+  /** 할인 비율 변경 */
+  const handleDiscountRateChange = useCallback(
+    (index: number, value: string) => {
+      const newDiscounts = [...productForm.discounts];
+      newDiscounts[index].rate = (parseInt(value) || 0) / 100;
+      updateField("discounts", newDiscounts);
+    },
+    [productForm.discounts, updateField]
+  );
+
+  /** 할인 제거 */
+  const handleRemoveDiscount = useCallback(
+    (index: number) => {
+      const newDiscounts = productForm.discounts.filter((_, i) => i !== index);
+      updateField("discounts", newDiscounts);
+    },
+    [productForm.discounts, updateField]
+  );
+
+  /** 할인 추가 */
+  const handleAddDiscount = useCallback(() => {
+    const newDiscount = DISCOUNT_OPTIONS.DEFAULT_DISCOUNT;
+    updateField("discounts", [...productForm.discounts, newDiscount]);
+  }, [productForm.discounts, updateField]);
+
+  /**상품 추가 취소 */
+  const handleCancel = useCallback(() => {
+    setEditingProduct(null);
+    setProductForm(INITIAL_PRODUCT_FORM);
+    setShowProductForm(false);
+  }, [setEditingProduct, setProductForm, setShowProductForm]);
+
+  return {
+    handleNameChange,
+    handleDescriptionChange,
+    handlePriceChange,
+    handlePriceBlur,
+    handleStockChange,
+    handleStockBlur,
+    handleDiscountQuantityChange,
+    handleDiscountRateChange,
+    handleRemoveDiscount,
+    handleAddDiscount,
+    handleCancel,
+  };
+};
